requirements.txt
build==1.0.3
CacheControl==0.13.1
certifi==2023.11.17
charset-normalizer==3.3.2
cleo==2.1.0
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.8
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.13.1
idna==3.6
installer==0.7.0
jaraco.classes==3.3.0
keyring==24.3.0
more-itertools==10.2.0
msgpack==1.0.7
packaging==23.2
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyproject_hooks==1.0.0
pywin32-ctypes==0.2.2
rapidfuzz==3.6.1
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.4
tomlkit==0.12.3
trove-classifiers==2024.1.8
urllib3==2.1.0
virtualenv==20.25.0
